Conflicts:
	src/main/java/egovframework/ezEKP/ezCommon/service/EzCommonService.java
	src/main/java/egovframework/ezEKP/ezNewPortal/dao/EzNewPortalDAO.java
	src/main/java/egovframework/ezEKP/ezNewPortal/service/EzNewPortalService.java
	src/main/java/egovframework/ezEKP/ezNewPortal/service/impl/EzNewPortalServiceImpl.java
	src/main/java/egovframework/ezEKP/ezNewPortal/web/EzNewPortalGWController.java
	src/main/resources/egovframework/sqlmap/ezEKP/ezCommon/EzCommon_SQL_Mysql.xml
	src/main/resources/egovframework/sqlmap/ezEKP/ezCommon/EzCommon_SQL_Oracle.xml
	src/main/resources/egovframework/sqlmap/ezEKP/ezNewPortal/EzNewPortal_SQL_Mysql.xml
	src/main/resources/egovframework/sqlmap/ezEKP/ezNewPortal/EzNewPortal_SQL_Oracle.xml
	src/main/resources/egovframework/sqlmap/ezEKP/init_data_mysql.sql
	src/main/resources/egovframework/sqlmap/ezEKP/init_data_oracle.sql
	src/main/webapp/js/ezNewPortal/lang/ezNewPortal_en.js
	src/main/webapp/js/ezNewPortal/lang/ezNewPortal_ja.js
	src/main/webapp/js/ezNewPortal/lang/ezNewPortal_ko.js

+++ src/main/java/egovframework/ezEKP/ezApprovalG/web/EzApprovalGController.java
@@ -931,6 +931,8 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		if (isUsed == null) {
 			isUsed = "";
 		}
+
+		String useOpenGov = config.getProperty("config.useOpenGov");
 		
 		String approvalFlag = ezCommonService.getTenantConfig("ApprovalFlag", userInfo.getTenantId());
 		String junGyulFlag = ezCommonService.getTenantConfig("JunGyulFlag", userInfo.getTenantId());
@@ -1039,6 +1041,10 @@ public class EzApprovalGController extends EgovFileMngUtil{
 			}
 		}
 		
+		//2020-01-28 김은석 추가
+		String useAnnualSusinYN = ezCommonService.getTenantConfig("useAnnualSusinYN", userInfo.getTenantId());
+		
+		model.addAttribute("useAnnualSusinYN", useAnnualSusinYN);
 		model.addAttribute("beforeDocID", beforeDocID);
 		model.addAttribute("isUsed", isUsed);
 		model.addAttribute("approvalFlag", approvalFlag);
@@ -1077,10 +1083,14 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		} else {
 			model.addAttribute("reformflag", ezApprovalGService.getReformInfoApprovalDocument(docID, userInfo.getId(), userInfo.getCompanyID(), tenantID).getReformFlag());
 		}
-		
-		String formId = ezApprovalGService.getFormId(formURL);
-		String openGovFlag = ezApprovalGService.getOpenGovFlag(formId, userInfo.getTenantId(), userInfo.getCompanyID());
-		model.addAttribute("openGovFlag", openGovFlag);
+
+		if (useOpenGov.equalsIgnoreCase("YES")) {
+            String formId = ezApprovalGService.getFormId(formURL);
+            String openGovFlag = ezApprovalGService.getOpenGovFlag(formId, userInfo.getTenantId(), userInfo.getCompanyID());
+            model.addAttribute("openGovFlag", openGovFlag);
+        }
+
+		model.addAttribute("useOpenGov", useOpenGov);
 		
 		logger.debug("draftui ended.");
 
@@ -1271,11 +1281,7 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		String companyID = userInfo.getCompanyID();
 		
 		String useOpenGov = config.getProperty("config.useOpenGov");
-		
-		if (useOpenGov != null && useOpenGov.equals("YES")) {
-			model.addAttribute("useOpenGov", useOpenGov);
-		}
-		
+
 		if (orgCompanyID != null && !orgCompanyID.equals("") && !orgCompanyID.equals(companyID)) {
 			userInfo.setCompanyID(orgCompanyID);
 		}
@@ -1345,7 +1351,8 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		model.addAttribute("orgCompanyID", orgCompanyID);
 		model.addAttribute("ext", ext);
 		model.addAttribute("useReceiveInfoName", useReceiveInfoName);
-		
+		model.addAttribute("useOpenGov", useOpenGov);
+
 		logger.debug("ezApprovalInfo ended.");
 		
 		return "ezApprovalG/apprGezApprovalInfo";
@@ -1662,9 +1669,14 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		userInfo = commonUtil.aprUserInfo(loginCookie);
 		
 		String docID = request.getParameter("docID");
+
+        //원문정보공개 데이터 제거
+        if (config.getProperty("useOpenGov").equalsIgnoreCase("YES")) {
+            ezApprovalGService.deleteOpenGovDocInfo(docID, userInfo.getCompanyID(), userInfo.getTenantId());
+        }
+
 		String result = ezApprovalGService.deleteDocInfo(docID, "CHECK", userInfo.getCompanyID(), userInfo.getTenantId());
-		//원문정보공개 데이터 제거
-		ezApprovalGService.deleteOpenGovDocInfo(docID, userInfo.getCompanyID(), userInfo.getTenantId());
+
 		logger.debug("undoDoc ended.");
 		
 		return result;
@@ -4089,6 +4101,8 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		} else {
 			susinAdmin = "NO";
 		}
+
+		String useOpenGov = config.getProperty("config.useOpenGov");
 		
 		String docID = request.getParameter("docID");
 		String uID = request.getParameter("id");
@@ -4213,7 +4227,11 @@ public class EzApprovalGController extends EgovFileMngUtil{
 			}
 		}
 		
-		model.addAttribute("optSignDateFormat", optSignDateFormat);
+		//2020-01-23 김은석 추가
+		String useAnnualSusinYN = ezCommonService.getTenantConfig("useAnnualSusinYN", userInfo.getTenantId());
+		
+		model.addAttribute("useAnnualSusinYN", useAnnualSusinYN);
+	    model.addAttribute("optSignDateFormat", optSignDateFormat);
 		model.addAttribute("optIsSplit", optIsSplit);
 		model.addAttribute("optSplitKind", optSplitKind);
 		model.addAttribute("optJunKyulInfo", optJunKyulInfo);
@@ -4249,8 +4267,7 @@ public class EzApprovalGController extends EgovFileMngUtil{
 			model.addAttribute("formId", formId);
 		}
 		
-		if (approvalFlag.equals("G")) {
-			//basis = "", reason = "", listOpenFlag = "", fileOpenFlagList = "";
+		if (useOpenGov.equalsIgnoreCase("YES") && approvalFlag.equalsIgnoreCase("G")) {
 			Map<String, Object> openGovMap = ezApprovalGService.getOpenGovInfo(docID, userInfo.getTenantId(), userInfo.getCompanyID());
 			
 			model.addAttribute("basis", openGovMap.get("basis"));
@@ -4258,6 +4275,8 @@ public class EzApprovalGController extends EgovFileMngUtil{
 			model.addAttribute("listOpenFlag", openGovMap.get("listOpenFlag"));
 			model.addAttribute("fileOpenFlagList", openGovMap.get("fileOpenFlagList"));
 		}
+
+		model.addAttribute("useOpenGov", useOpenGov);
 		
 		logger.debug("approvui ended");
 		
@@ -5150,6 +5169,7 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		model.addAttribute("isNonElecRec", isNonElecRec);
 		model.addAttribute("useReceiveDocNo", useReceiveDocNo);
 		model.addAttribute("docNumZeroCnt", Integer.parseInt(docNumZeroCnt));
+		model.addAttribute("useOpenGov", config.getProperty("useOpenGov"));
 
 		logger.debug("recevGSusin ended.");
 		
@@ -6471,9 +6491,9 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		
 		String result = "";
 	
-		docNumber = xmlDom.getDocumentElement().getChildNodes().item(0).getTextContent().replace("[", "[[]").replace("%", "[%]").replace("_", "[_]");
-		docTitle = xmlDom.getDocumentElement().getChildNodes().item(1).getTextContent().replace("[", "[[]").replace("%", "[%]").replace("_", "[_]");
-        drafter = xmlDom.getDocumentElement().getChildNodes().item(2).getTextContent().replace("[", "[[]").replace("%", "[%]").replace("_", "[_]");
+		docNumber = xmlDom.getDocumentElement().getChildNodes().item(0).getTextContent().replace("[", "\\[").replace("%", "\\%").replace("_", "\\_");
+		docTitle = xmlDom.getDocumentElement().getChildNodes().item(1).getTextContent().replace("[", "\\[").replace("%", "\\%").replace("_", "\\_");
+        drafter = xmlDom.getDocumentElement().getChildNodes().item(2).getTextContent().replace("[", "\\[").replace("%", "\\%").replace("_", "\\_");
         String draftfrom = xmlDom.getDocumentElement().getChildNodes().item(3).getTextContent();
         String draftto = xmlDom.getDocumentElement().getChildNodes().item(4).getTextContent();
         String apprfrom = xmlDom.getDocumentElement().getChildNodes().item(5).getTextContent();
@@ -6481,7 +6501,7 @@ public class EzApprovalGController extends EgovFileMngUtil{
         String mypapprfrom = xmlDom.getDocumentElement().getChildNodes().item(7).getTextContent();
         String mypapprto = xmlDom.getDocumentElement().getChildNodes().item(8).getTextContent();
         formID = xmlDom.getDocumentElement().getChildNodes().item(9).getTextContent();
-        draftDeptName = xmlDom.getDocumentElement().getChildNodes().item(11).getTextContent().replace("[", "[[]").replace("%", "[%]").replace("_", "[_]");
+        draftDeptName = xmlDom.getDocumentElement().getChildNodes().item(11).getTextContent().replace("[", "\\[").replace("%", "\\%").replace("_", "\\_");
         containerID = xmlDom.getDocumentElement().getChildNodes().item(12).getTextContent();
         userID = xmlDom.getDocumentElement().getChildNodes().item(13).getTextContent();
         pageNum = xmlDom.getDocumentElement().getChildNodes().item(16).getTextContent();
@@ -9243,6 +9263,12 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		userInfo = commonUtil.userInfo(loginCookie);
 		String docID = request.getParameter("docID");
 		String companyID = userInfo.getCompanyID();
+		String orgCompanyID = request.getParameter("orgCompanyID");
+		
+		if (orgCompanyID != null && !orgCompanyID.equals("") && !orgCompanyID.equals(userInfo.getCompanyID())) {
+			companyID = orgCompanyID;
+		}
+		
 		int tenantID = userInfo.getTenantId();
 		
 		String ext = ezApprovalGService.getDocExt(docID, companyID, tenantID);
@@ -9503,6 +9529,7 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		model.addAttribute("useReceiveDocNo", useReceiveDocNo);
 		model.addAttribute("docNumZeroCnt", docNumZeroCnt);
 		model.addAttribute("isReDraft", isReDraft);
+		model.addAttribute("useOpenGov", config.getProperty("config.useOpenGov"));
 		
 		logger.debug("recevGDeptHapyui ended");
 
@@ -9747,13 +9774,9 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		
 		JSONObject openGovJson = new JSONObject();
 		openGovJson.put("openGovInfo", openGovInfo);
-//			attachJson.put("fileOpenFlag", attach.getFileOpenFlag());
-//			attachJson.put("sn", attach.getAttachFileSN());
-//			attachJson.put("fileName", attach.getAttachFileName());
-//			attachJson.put("fileSize", fileSize);
-//			
 		
 		logger.debug("getOpenGovInfoForUpdate ended.");
+
 		return openGovJson;
 	}
 
@@ -9818,10 +9841,6 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		openGovJson.put("reason", openGovMap.get("reason"));
 		openGovJson.put("listOpenFlag", openGovMap.get("listOpenFlag"));
 		openGovJson.put("fileOpenFlagList", openGovMap.get("fileOpenFlagList"));
-		/*attachJson.put("fileOpenFlag", attach.getFileOpenFlag());
-		attachJson.put("sn", attach.getAttachFileSN());
-		attachJson.put("fileName", attach.getAttachFileName());
-		attachJson.put("fileSize", fileSize);*/
 
 		logger.debug("getOpenGovDocInfo ended.");
